refactor: add type in types filenames
