+++ src/shared/types/coordinates.type.ts
