+++ src/shared/types/index.ts
@@ -1,5 +1,5 @@
-export * from './comment.js';
-export * from './offer.js';
-export * from './user.js';
-export * from './coordinates.js';
+export * from './comment.type.js';
+export * from './offer.type.js';
+export * from './user.type.js';
+export * from './coordinates.type.js';
 export * from './mock-server-data.type.js';
