+++ src/shared/types/user.type.ts
